added move() method for ball, and changed types to double

+++ src/model/Vector.java
@@ -2,9 +2,9 @@ package model;
 
 public class Vector {
     // simple class that saves a 2-dimensional position
-    int x;
-    int y;
-    Vector(int x, int y) {
+    double x;
+    double y;
+    Vector(double x, double y) {
         this.x = x;
         this.y = y;
     }
